[REFACTOR] (Output) : output class 통합
- 여러개의 함수로 나뉘어져있던 output 함수를 Output class로 통합
- output 이라는 폴더명을 src 바로 아래로 수정
- 함수 앞에 붙어있던 'print' 명칭을 제거
(Output class 안에 속해있기 때문에 호출할때도 Output.winner 과 같은 식으로 호출됨)
(Output.printWinner 의 경우엔 출력의 의미가 2번 중복된다는 느낌을 받아 수정함)

여러개의 파일로 나뉘어져있던 output 함수들을 하나의 Output class로 통합하여 사용하였습니다
