+++ src/App.js
@@ -1,6 +1,6 @@
-import printWinner from "./Components/Output/printWinner.js";
 import raceCar from "./Components/RaceCar/raceCar.js";
 import Input from "./Input/Input.js";
+import Output from "./Output/Output.js";
 
 class App {
   async run() {
@@ -8,7 +8,7 @@ class App {
       const carList = await Input.CarList();
       const tryNumber = await Input.TryNum();
       const winnerCars = raceCar(carList, tryNumber);
-      printWinner(winnerCars);
+      Output.winner(winnerCars);
     } catch (error) {
       throw new Error(error.message);
     }
