--- src/Components/Output/printResultMessage.js
@@ -1,11 +1,0 @@
-import { Console } from "@woowacourse/mission-utils";
-
-/**
- *
- * @param {string} resultMessage
- * @summary 결과 메시지를 출력한다.
- */
-const printResultMessage = (resultMessage) => {
-  Console.print(resultMessage);
-};
-export default printResultMessage;
